Handle enter on PF3 submit wizard and share code
